pyproject.toml
[project]
name = "account-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "celery>=5.5.3",
    "cryptography>=45.0.3",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "djangorestframework>=0.0.1",
    "djangorestframework-simplejwt>=5.5.0",
    "isort>=6.0.1",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
    "isort>=6.0.1",
]

[tool.setuptools]
packages = ["accounts", "users", "config"]

[tool.isort]
profile = "black"

